Spit `DictionaryPipelineExecutor` into separate file
