+++ src/Dictionaries/registerRangeHashedDictionary.cpp
@@ -1,5 +1,8 @@
-#include "RangeHashedDictionary.h"
+#include <Dictionaries/RangeHashedDictionary.h>
+
 #include <Dictionaries/DictionarySource.h>
+#include <Dictionaries/ClickHouseDictionarySource.h>
+#include <Dictionaries/DictionarySourceHelpers.h>
 #include <Dictionaries/DictionaryFactory.h>
 
 namespace DB
